+++ eggs/Naaya/eggs/Naaya/Products/Naaya/managers/skel_parser.py
@@ -1,4 +1,3 @@
-# Python imports
 from xml.dom import minidom
 
 # Translation Map