+++ eggs/Naaya/eggs/Naaya/Products/NaayaCore/FormsTool/interfaces.py
@@ -1,0 +1,5 @@
+from zope.interface import Interface
+
+class ITemplate(Interface):
+    def __call__(*args, **kwargs):
+        """Render the template"""