+++ eggs/Naaya/eggs/Naaya/naaya/content/base/meta.py
@@ -1,25 +1,19 @@
-# Python imports
 import os
-from copy import copy
-import types
 
-# Zope imports
+import zLOG
 from AccessControl import ClassSecurityInfo
 from Globals import InitializeClass
 from OFS.Application import Application
 from OFS.misc_ import Misc_
-import zLOG
 from zope.interface import Interface
 from zope.interface import implements
 from zope.configuration.fields import GlobalObject
 from zope.app.i18n import ZopeMessageFactory as _
 
-# Products imports
 from constants import *
-from Products.NaayaBase.NyValidation import NyValidation
 from Products.Naaya.NyFolderBase import NyFolderBase
 from Products.NaayaCore.SchemaTool.SchemaTool import lookup_schema_product
-
+from Products.NaayaCore.FormsTool.NaayaTemplate import NaayaPageTemplateFile
 
 class INaayaContent(Interface):
     """ Basic Naaya Content """
@@ -59,24 +53,31 @@ class NaayaContent(object):
         return self._contents.get(key)
 
 def register_naaya_content(_context, factory, **kwargs):
-    """ """
+    """ Register a naaya content type.
+
+    This is called for each content type using it's config dictionary. This will
+    validate the config of the content type, perform a few sanity checks,
+    declare the permissions for the content type and register the templates.
+
+    """
+
+    assert factory, "No factory provided"
+    assert callable(factory), "Factory must be callable"
+    config = factory()
+
+    assert 'meta_type' in config, "Config must contain at least a meta_type"
+
     _contents = NaayaContent._contents
     _misc = NaayaContent._misc
     _constants = NaayaContent._constants
 
-    if not factory:
-        raise TypeError("No factory provided")
-
-    if isinstance(factory, types.FunctionType):
-        factory = factory()
-    else:
-        return
-    _contents[factory['meta_type']] = factory
-    _misc.update(factory.get('_misc', {}))
+    _contents[config['meta_type']] = config
+    _misc.update(config.get('_misc', {}))
 
     # register _misc into Zope
     if not Application.misc_.__dict__.has_key('NaayaContent'):
-        Application.misc_.__dict__['NaayaContent'] = Misc_('NaayaContent', _misc)
+        Application.misc_.__dict__['NaayaContent'] = Misc_('NaayaContent',
+                _misc)
     else:
         Application.misc_.__dict__['NaayaContent'].__dict__['_d'].update(_misc)
 
@@ -86,15 +87,22 @@ def register_naaya_content(_context, factory, **kwargs):
     security = ClassSecurityInfo()
     NyFolderBase.security = security
 
-    if factory['meta_type'] != 'Naaya Folder' and 'folder_constructors' in factory:
-
-        for folder_property, object_constructor in factory['folder_constructors']:
+    # declare permissions for object constructors
+    if (config['meta_type'] != 'Naaya Folder' and
+        'folder_constructors' in config):
+        for folder_property, object_constructor in config['folder_constructors']:
             setattr(NyFolderBase, folder_property, object_constructor)
-            NyFolderBase.security.declareProtected(factory['permission'], folder_property)
+            NyFolderBase.security.declareProtected(config['permission'],
+                    folder_property)
+    # register forms
+    for form in config.get('forms', []):
+        NaayaPageTemplateFile(os.path.join(config['package_path'], 'zpt', form),
+                globals(), form)
 
     InitializeClass(NyFolderBase)
     # log success
-    zLOG.LOG('naaya.content', zLOG.DEBUG, 'Pluggable module "%s" registered' % factory['meta_type'])
+    zLOG.LOG('naaya.content', zLOG.DEBUG,
+            'Pluggable content type "%s" registered' % config['meta_type'])
 
 def get_schema_name(portal, meta_type):
     if meta_type == 'Naaya Folder':