+++ eggs/Naaya/eggs/Naaya/CHANGELOG.rst
@@ -3,6 +3,9 @@
 * refactor: :mod:`Products.Naaya.NySite` stores Zope and
   Naaya containers meta_type-s in two lists in the beginning of the module
 * refactor: :mod:`naaya.i18n` replaces Localizer and itools
+* refactor: :mod:`Products.NaayaCore.FormsTool` templates registered via ZCA
+  and bundles [plugaale, andredor, moregale]
+  update script: "Migrate to bundles"
 
 2.11.5 (2011-09-23)
 --------------------